requirements.txt
# LangGraph for agent framework
langgraph>=0.2.0
langgraph-checkpoint

# LangChain for Google Gemini
langchain-google-genai>=0.0.5
langchain-core>=0.3.0

# Google Gemini SDK
google-generativeai>=0.8.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
